# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ActionStatusUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/ActionItemStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "ActionStatusUpdate",
        "type": "object"
      },
      "ActionView": {
        "properties": {
          "action_type": {
            "$ref": "#/components/schemas/ActionItemType"
          },
          "addressee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Addressee"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "due_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "status": {
            "$ref": "#/components/schemas/ActionItemStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "due_date",
          "action_type",
          "status",
          "location",
          "addressee"
        ],
        "title": "ActionView",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Sanctuary API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/action-items/{item_id}": {
      "patch": {
        "operationId": "v1_set_action_status_api_v1_action_items__item_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Set Action Status",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/cases/{case_id}/confirm-draft": {
      "post": {
        "description": "Ratify an AI-created draft case.",
        "operationId": "v1_confirm_draft_api_v1_cases__case_id__confirm_draft_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseRef"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Confirm Draft",
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/cases/{case_id}/reject-draft": {
      "post": {
        "description": "Delete an AI-created draft case; its documents return to triage.",
        "operationId": "v1_reject_draft_api_v1_cases__case_id__reject_draft_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Reject Draft",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{doc_id}": {
      "delete": {
        "operationId": "v1_delete_document_api_v1_documents__doc_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Delete Document",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{doc_id}/cost-signals/promote": {
      "post": {
        "description": "Create a LegalCost from the document's most recent cost signal.\n\nInvoice/advance signals already materialise during enrichment; this is\nfor the remaining kinds (e.g. a Streitwert ruling) with optional overrides.",
        "operationId": "v1_promote_cost_signal_api_v1_documents__doc_id__cost_signals_promote_post",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CostPromotion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostPromoted"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Promote Cost Signal",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{doc_id}/metadata": {
      "put": {
        "description": "Apply a metadata patch (fields left null are untouched) and recompute review flags.",
        "operationId": "v1_update_metadata_api_v1_documents__doc_id__metadata_put",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MetadataUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentReview"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Update Metadata",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{doc_id}/pipeline/retry-all": {
      "post": {
        "operationId": "v1_retry_all_stages_api_v1_documents__doc_id__pipeline_retry_all_post",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineView"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Retry All Stages",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{doc_id}/pipeline/{stage}/retry": {
      "post": {
        "operationId": "v1_retry_stage_api_v1_documents__doc_id__pipeline__stage__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "stage",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PipelineStage"
            }
          },
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineView"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Retry Stage",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/documents/{doc_id}/reactions": {
      "post": {
        "description": "Toggle a reaction; with ``notes`` it is upserted instead of toggled.",
        "operationId": "v1_set_reaction_api_v1_documents__doc_id__reactions_post",
        "parameters": [
          {
            "in": "path",
            "name": "doc_id",
            "required": true,
            "schema": {
              "title": "Doc Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReactionView"
                  },
                  "title": "Response V1 Set Reaction Api V1 Documents  Doc Id  Reactions Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "V1 Set Reaction",
        "tags": [
          "documents"
        ]
      }
    },
...
  }
}